refactor: Extract submission title to its own component
